Tensor:cdiv with three arguments
Torch tensors support the cdiv operation with both two and three arguments.
cutorch only supports the two-argument (in-place) version. This patch extends it
to also support the three-argument version and adds a test for it. The
implementation is very similar to cmul which already supports both two and three
arguments.