Add batch matrix-matrix multiplication.
Add torch.bmm, which expects two 3D tensors as inputs. A 'b x m x n'
tensor is interpreted as a batch of b matrices of size 'm x n'.
The result is a batch of matrices computed by multiplying the
corresponding pair of matrices in the two input batches.